Pr 반영

+++ src/main/java/com/beauty/api/model/user/service/MemberService.java
@@ -1,9 +1,11 @@
 package com.beauty.api.model.user.service;
 
-import com.beauty.api.model.user.dto.Member;
+import com.beauty.api.model.user.domain.Member;
+import com.beauty.api.model.user.dto.MemberFindEmail;
 import com.beauty.api.model.user.dto.MemberResponse;
 import com.beauty.api.model.user.dto.MemberSignInRequest;
 import com.beauty.api.model.user.dto.MemberSignUpRequest;
+import com.beauty.api.model.user.dto.MemberUpdatePassword;
 import com.beauty.api.model.user.dto.MemberUpdateRequest;
 import com.beauty.api.model.user.persist.entity.MemberEntity;
 import com.beauty.api.model.user.persist.repository.MemberRepository;
@@ -31,13 +33,14 @@ public class MemberService implements UserDetailsService {
   public MemberResponse signUp(MemberSignUpRequest memberSignUpRequest) {
     boolean exists = this.memberRepository.existsByEmail(memberSignUpRequest.getEmail());
     if (exists) {
-      throw new IllegalArgumentException("이미 존재하는 아이디 입니다.");
+      throw new IllegalArgumentException("이미 가입되어 있는 이메일 입니다.");
     }
 
     memberSignUpRequest.setPassword(this.passwordEncoder.encode(memberSignUpRequest.getPassword()));
-    MemberEntity result = this.memberRepository.save(memberSignUpRequest.toEntity());
 
-    return MemberResponse.fromEntity(result);
+    Member member = Member.fromRequest(memberSignUpRequest);
+
+    return MemberResponse.fromEntity(this.memberRepository.save(member.toEntity()));
   }
 
   public MemberEntity signIn(MemberSignInRequest memberSigninReQuest) {
@@ -51,19 +54,57 @@ public class MemberService implements UserDetailsService {
     return memberEntity;
   }
 
-  public MemberResponse updateUser(MemberUpdateRequest memberUpdateRequest) {
-    MemberEntity memberEntity = this.memberRepository.findByEmail(memberUpdateRequest.getEmail())
+  public MemberResponse updateMember(Member member, MemberUpdateRequest memberUpdateRequest) {
+    MemberEntity memberEntity = this.memberRepository.findById(memberUpdateRequest.getId())
         .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 아이디 입니다."));
 
-    this.memberRepository.save(memberUpdateRequest.toEntity());
+    if (!memberEntity.getId().equals(member.getId())) {
+      throw new IllegalArgumentException("해당 회원의 정보가 아닙니다.");
+    }
 
-    return MemberResponse.fromEntity(memberEntity);
+    return MemberResponse.fromEntity(this.memberRepository.save(member.update(memberUpdateRequest).toEntity()));
   }
 
-  public void deleteUser(Long id) {
+  public void deleteMember(Member member, Long id) {
     MemberEntity memberEntity = this.memberRepository.findById(id)
         .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 아이디 입니다."));
 
+    if (!member.getId().equals(id)) {
+      throw new IllegalArgumentException("해당 회원의 정보가 아닙니다.");
+    }
+
     this.memberRepository.delete(memberEntity);
+  }
+
+  public MemberResponse getMember(Member member, Long id) {
+    MemberEntity memberEntity = this.memberRepository.findById(id)
+        .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 아이디 입니다."));
+
+    if (!member.getId().equals(id)) {
+      throw new IllegalArgumentException("해당 회원의 정보가 아닙니다.");
+    }
+    return MemberResponse.fromEntity(memberEntity);
+  }
+
+  public MemberResponse updatePassword(Member member, MemberUpdatePassword memberUpdatePassword) {
+    MemberEntity memberEntity = this.memberRepository.findById(memberUpdatePassword.getId())
+        .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 아이디 입니다."));
+
+    if (!memberEntity.getId().equals(member.getId())) {
+      throw new IllegalArgumentException("해당 회원의 정보가 아닙니다.");
+    }
+
+    if (!this.passwordEncoder.matches(memberUpdatePassword.getPassword(), memberEntity.getPassword())) {
+      throw new IllegalArgumentException("비밀번호가 일치하지 않습니다.");
+    }
+    member.setPassword(this.passwordEncoder.encode(memberUpdatePassword.getNewPassword()));
+    return MemberResponse.fromEntity(this.memberRepository.save(member.toEntity()));
+  }
+
+  public MemberResponse findEmail(MemberFindEmail memberFindEmail) {
+    MemberEntity memberEntity = this.memberRepository.findByNameAndPhone(memberFindEmail.getName(),
+        memberFindEmail.getPhone()).orElseThrow(() -> new IllegalArgumentException("존재하지 않는 회원입니다."));
+
+    return MemberResponse.fromEntity(memberEntity);
   }
 }
